Feat: Enable sticky header on stats table
Implements a sticky header for the student stats table by adding a fixed height and vertical overflow to its container.

`To achieve this, the following properties were added to the #student-stats-table-container in style.css:

overflow-y: auto

max-height: 50vh`

This change introduces a vertical scrollbar within the table container, allowing the <thead> to remain sticky while the <tbody> scrolls.