Add assessment mode for student selection
Introduces an assessment mode that changes the logic for selecting students, allowing for a separate workflow when grading. Adds toggling of assessment mode via long-press on the select button, visual feedback, and exposes internal modules in developer mode. Updates state management to track assessment mode and provides related UI notifications.

Introduces a dev mode in the console by clicking 10 times on the app title on the main page. This dev tool exposes all the major files of the project to the console for developing purposes